// ArchiCore/src/app/api-client/model/missionResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MissionResponse { 
    codeMission?: string;
    exercice?: string;
    codeDossier?: string;
    statut?: MissionResponse.StatutEnum;
    createdAt?: string;
}
export namespace MissionResponse {
    export const StatutEnum = {
        Planifiee: 'PLANIFIEE',
        EnCours: 'EN_COURS',
        Terminee: 'TERMINEE',
        Suspendue: 'SUSPENDUE'
    } as const;
    export type StatutEnum = typeof StatutEnum[keyof typeof StatutEnum];
}